Re-work the CDM example.
Previously, it was writing contracts to the ledger. We've removed this as it
would have required distributing the CDM data type templates *.dar file too.
For versioning etc. this would not have been a good idea, and would also
confuse the participants if we had two sets of cdm sources!